+++ Sims1WidescreenPatcher.Core/Factories/UserControlViewModelCreator.cs
@@ -1,0 +1,8 @@
+﻿using ReactiveUI;
+
+namespace Sims1WidescreenPatcher.Core.Factories;
+
+public abstract class UserControlViewModelCreator
+{
+    public abstract IReactiveObject FactoryMethod(string arg0);
+}
