+++ Sims1WidescreenPatcher.Core/Tabs/IOptionalTabViewModel.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel;
+using ReactiveUI;
+
+namespace Sims1WidescreenPatcher.Core.Tabs;
+
+public interface IOptionalTabViewModel
+{
+    IDisposable SuppressChangeNotifications();
+    bool AreChangeNotificationsEnabled();
+    IDisposable DelayChangeNotifications();
+    IObservable<IReactivePropertyChangedEventArgs<IReactiveObject>> Changing { get; }
+    IObservable<IReactivePropertyChangedEventArgs<IReactiveObject>> Changed { get; }
+    IObservable<Exception> ThrownExceptions { get; }
+    event PropertyChangingEventHandler? PropertyChanging;
+    event PropertyChangedEventHandler? PropertyChanged;
+}
