+++ Sims1WidescreenPatcher.Core/ViewModels/ICheckboxViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Sims1WidescreenPatcher.Core.ViewModels;
+
+public interface ICheckboxViewModel
+{
+    string Label { get; }
+    bool Checked { get; set; }
+}
