+++ Sims1WidescreenPatcher.Core/ViewModels/ViewModelBase.cs
@@ -4,4 +4,5 @@ namespace Sims1WidescreenPatcher.Core.ViewModels;
 
 public class ViewModelBase : ReactiveObject
 {
+    
 }
