+++ Sims1WidescreenPatcher.UI/App.axaml
@@ -1,6 +1,5 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:vl="using:Sims1WidescreenPatcher.UI"
              x:Class="Sims1WidescreenPatcher.UI.App"
 			 RequestedThemeVariant="Default">
