+++ Sims1WidescreenPatcher.UI/Dialogs/CustomInformationDialog.axaml
@@ -4,7 +4,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:vm="clr-namespace:Sims1WidescreenPatcher.Core.ViewModels;assembly=Sims1WidescreenPatcher.Core"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="Sims1WidescreenPatcher.UI.Views.CustomInformationDialog"
+        x:Class="Sims1WidescreenPatcher.UI.Dialogs.CustomInformationDialog"
         Icon="/Assets/SimsICO.ico"
         MinWidth="300"
         MinHeight="100"
