+++ Sims1WidescreenPatcher.UI/Dialogs/CustomInformationDialog.axaml.cs
@@ -2,7 +2,7 @@
 using Avalonia.ReactiveUI;
 using Sims1WidescreenPatcher.Core.ViewModels;
 
-namespace Sims1WidescreenPatcher.UI.Views;
+namespace Sims1WidescreenPatcher.UI.Dialogs;
 
 public partial class CustomInformationDialog : ReactiveWindow<CustomInformationDialogViewModel>
 {
