+++ Sims1WidescreenPatcher.UI/Styles/TabStyles.axaml
@@ -1,0 +1,18 @@
+﻿<Styles xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <Design.PreviewWith>
+        <Border Padding="20">
+            <!-- Add Controls for Previewer Here -->
+        </Border>
+    </Design.PreviewWith>
+
+    <Style Selector="TabControl.tabcontrol">
+        <Setter Property="Padding" Value="0" />
+    </Style>
+    
+    <Style Selector="TabItem.tabitem">
+        <Setter Property="FontSize" Value="16" />
+        <Setter Property="MinHeight" Value="24" />
+        <Setter Property="MaxHeight" Value="24" />
+    </Style>
+</Styles>
