+++ Sims1WidescreenPatcher/DependencyInjection/Bootstrapper.cs
@@ -7,6 +7,7 @@ namespace Sims1WidescreenPatcher.DependencyInjection
 		public static void Register(IMutableDependencyResolver services, IReadonlyDependencyResolver resolver)
 		{
 			ServicesBootstrapper.RegisterServices(services, resolver);
+			FactoryBootstrapper.RegisterFactories(services, resolver);
 			ViewModelBootstrapper.RegisterViewModels(services, resolver);
 			ViewBootstrapper.RegisterViews(services, resolver);
 		}
