+++ Sims1WidescreenPatcher/DependencyInjection/FactoryBootstrapper.cs
@@ -1,0 +1,12 @@
+﻿using Sims1WidescreenPatcher.Core.Factories;
+using Splat;
+
+namespace Sims1WidescreenPatcher.DependencyInjection;
+
+public static class FactoryBootstrapper
+{
+    public static void RegisterFactories(IMutableDependencyResolver services, IReadonlyDependencyResolver resolver)
+    {
+        services.Register(() => new CheckboxViewModelFactory());
+    }
+}
