+++ Sims1WidescreenPatcher/DependencyInjection/ViewBootstrapper.cs
@@ -1,4 +1,6 @@
-﻿using Sims1WidescreenPatcher.Core.ViewModels;
+﻿using Sims1WidescreenPatcher.Core.Tabs;
+using Sims1WidescreenPatcher.Core.ViewModels;
+using Sims1WidescreenPatcher.UI.Tabs;
 using Sims1WidescreenPatcher.UI.Views;
 using Splat;
 
@@ -11,5 +13,8 @@ public static class ViewBootstrapper
         services.RegisterLazySingleton(() => 
                 new MainWindow(resolver.GetService<IMainWindowViewModel>() 
                                ?? throw new InvalidOperationException()));
+        services.RegisterLazySingleton(() =>
+            new MainTab(resolver.GetService<IMainTabViewModel>() ?? throw new InvalidOperationException()));
+        services.RegisterLazySingleton(() => new OptionalTab());
     }
 }
