+++ Sims1WidescreenPatcher/DependencyInjection/ViewModelBootstrapper.cs
@@ -1,4 +1,6 @@
-﻿using Sims1WidescreenPatcher.Core.Services;
+﻿using Sims1WidescreenPatcher.Core.Factories;
+using Sims1WidescreenPatcher.Core.Services;
+using Sims1WidescreenPatcher.Core.Tabs;
 using Sims1WidescreenPatcher.Core.ViewModels;
 using Sims1WidescreenPatcher.Utilities.Services;
 using Splat;
@@ -11,13 +13,17 @@ namespace Sims1WidescreenPatcher.DependencyInjection
         {
             services.Register(() => new CustomResolutionDialogViewModel(), typeof(ICustomResolutionDialogViewModel));
             services.Register(() => new CustomYesNoDialogViewModel());
-            services.Register(() => new MainWindowViewModel(
+            services.Register(() => new MainWindowViewModel(), typeof(IMainWindowViewModel));
+            services.Register(() => new MainTabViewModel(
                 resolver.GetService<IResolutionsService>() ?? throw new InvalidOperationException(),
                 resolver.GetService<CustomYesNoDialogViewModel>() ?? throw new InvalidOperationException(),
                 resolver.GetService<ICustomResolutionDialogViewModel>() ?? throw new InvalidOperationException(),
                 resolver.GetService<IProgressService>() ?? throw new InvalidOperationException(),
                 resolver.GetService<IFindSimsPathService>() ?? throw new InvalidOperationException()),
-                typeof(IMainWindowViewModel));
+                typeof(IMainTabViewModel));
+            services.Register(() => new OptionalTabViewModel(
+                resolver.GetService<CheckboxViewModelFactory>() ?? throw new InvalidOperationException()),
+                typeof(IOptionalTabViewModel));
         }
     }
 }
